WIP: refactored a bit topics-related code
